PP-4936 Whitelist rules for body of Smartpay notifications
Whitelist the following rules for the body of a Smartpay notification.
1009 - "="
1010 - "("
1011 - ")"
1012 - "\'"

Smartpay notifications come from a trusted URL and we do secret
validation on incoming requests to ensure this. Therefore, if the body
is well-formed JSON - which is enforced by NAXSI the risk is very low
of having any malicious content in the notification.

It appears that we are receiving Smartpay notifications containing
fields that we weren't previously receiving which is why we are having
to loosen the NAXSI rules.

with @heathd