PP-4963 Whitelist rules for Stripe notification body
Whitelist the following rules for the body of a Stripe notification:
1010 - "("
1011 - ")"
1013 - "\'"
1015 - ","
1314 - "`"

Stripe notifications come from a trusted URL and we do secret
validation on incoming requests to ensure this. Therefore, if the body
is well-formed JSON, which is enforced by NAXSI, the risk is very low
of having any malicious content in the notification.