PP-5843 Whitelist comma in body for smartpay notifications
Rule 1015 is a comma

As notifications come from a whitelisted IP, we generally unblock rules
we're blocking on if it seems sensible that the notification could
contain the character we’re blocking on. In this case it does as the
field is 'operations' which suggests the value could be a comma
separated list.