Zen desk 3954915 Whitelist smartpay and stripe notifications
Whitelist rule 1000 for smartpay which restricted sql keywords and was
blocking the `authcode` in the body.

Whitelist rules for body of smartpay to align with address fields in
frontend.

Whitelist rule 1001 for stripe which restricted double quotes and was
blocking on the `line1` field in the body.

zendesk ticket https://govuk.zendesk.com/agent/tickets/3954915