PP-7384: Stop propagating GoCardless traffic to direct debit connector app
We are going to be removing IP firewall rules for notifications from the
payment gateways and Jenkins nodes. The firewall rules will be applied
exclusively in the connector app. However there are no rules for the direct
debit connector app as we don't expect any traffic from GoCardless. This PR
ensures that no traffic gets to the direct debit connector app.